+++ src/main/webapp/WEB-INF/faces-config.xml
@@ -4,5 +4,17 @@
               xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
                                   http://xmlns.jcp.org/xml/ns/javaee/web-facesconfig_2_2.xsd">
 
-
+    <navigation-rule>
+        <from-view-id>../views/login.xhtml</from-view-id>
+        <navigation-case>
+            <from-action>#{userConnectionBean.connectionLogin}</from-action>
+            <from-outcome>success</from-outcome>
+            <to-view-id>../views/index.xhtml</to-view-id>
+        </navigation-case>
+        <navigation-case>
+            <from-action>#{userConnectionBean.connectionLogin}</from-action>
+            <from-outcome>fail</from-outcome>
+            <to-view-id>../views/login.xhtml</to-view-id>
+        </navigation-case>
+    </navigation-rule>
 </faces-config>
