+++ src/main/webapp/views/login.xhtml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+        "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://xmlns.jcp.org/jsf/html"
+      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+      xmlns:f="http://xmlns.jcp.org/jsf/core">
+<f:view>
+    <h:outputLabel value="Login"/>
+    <h:form>
+        <h:inputText id="loginInput" value="#{userConnectionBean.login}" label="Lgin">
+        </h:inputText>
+        <h:inputSecret id="passwordInput" value="#{userConnectionBean.password}">
+<!--            <f:validateRegex pattern=""/>-->
+        </h:inputSecret>
+        <h:commandButton value="Submit" action="#{userConnectionBean.connectionLogin}"/>
+        <h:message for="passwordInput" style="color: darkred"/>
+    </h:form>
+</f:view>
+</html>
