+++ src/main/webapp/views/templates/header.xhtml
@@ -15,7 +15,8 @@
                 </div>
                 <h:outputText value=" Rent School 2022" styleClass="h2"></h:outputText>
             </div>
-            <h:commandButton value="Login" styleClass="btn btn-secondary my-2 my-sm-0" type="submit"/>
+            <h:outputText value="Welcome, #{userConnectionBean.login} "/>
+            <h:commandButton value="Disconect" styleClass="btn btn-secondary my-2 my-sm-0" type="submit"/>
         </div>
     </ui:composition>
     </body>
