+++ src/main/webapp/views/templates/mainPage.xhtml
@@ -16,7 +16,9 @@
    <ui:insert name="menu">
       <ui:include src="menu.xhtml"></ui:include>
    </ui:insert>
-   <ui:insert name="content"></ui:insert>
+   <ui:insert name="content">
+
+   </ui:insert>
    <ui:insert name="footer">
       <ui:include src="footer.xhtml"></ui:include>
    </ui:insert>
